Move indent_and_wrap into messages module.
